// x-weld-app-older-version/src/store/config/types.ts
export interface HostConfig {
  endpoints: string[];
  blacklist: string[];
  hosted: boolean;
  themePresets: SupportedTheme[];
}
// ...
export interface SupportedTheme {
  name: string;
  logo: SupportedThemeLogo;
  color: string;
  isDark: boolean;
}
// ...
export interface InitConfig {
  apiConfig: ApiConfig | InstanceConfig;
  hostConfig?: HostConfig;
  apiConnected?: boolean;
  apiAuthenticated?: boolean;
}

export interface ApiConfig {
  apiUrl: string;
  socketUrl: string;
}

// Saved to localstorage.
export interface InstanceConfig extends ApiConfig {
  name: string;
  active: boolean;
}